Prosseguindo com mapeamento de entidades

+++ mage/src/main/java/MAGE/mage/repository/MaquinaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MaquinaRepository extends JpaRepository<Maquina, Long> {
+public interface MaquinaRepository extends JpaRepository<Maquina, Integer> {
 }
